// src/shared/models/v1/ogame/buildings/BuildingType.ts
export enum BuildingType {
    metalMine = 1,
    crystalMine = 2,
    deuteriumSynthesizer = 3,

    metalStorage = 22,
    crystalStorage = 23,
    deuteriumTank = 24,

    solarPlant = 4,
    fusionReactor = 12,

    roboticsFactory = 14,
    shipyard = 21,
    researchLab = 31,
    allianceDepot = 34,
    missileSilo = 44,
    naniteFactory = 15,
    terraformer = 33,
    spaceDock = 36,

    lunarBase = 41,
    sensorPhalanx = 42,
    jumpGate = 43,
}