// src/shared/models/v1/ogame/ships/ShipType.ts
export enum ShipType {
    smallCargo = 202,
    largeCargo = 203,
    lightFighter = 204,
    heavyFighter = 205,
    cruiser = 206,
    battleship = 207,
    colonyShip = 208,
    recycler = 209,
    espionageProbe = 210,
    bomber = 211,
    solarSatellite = 212,
    destroyer = 213,
    deathStar = 214,
    battlecruiser = 215,
    crawler = 217,
    reaper = 218,
    pathfinder = 219,
}